requirements.txt
# 核心依赖
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# 技术指标
ta-lib>=0.4.0

# 数据可视化
matplotlib>=3.7.0
plotly>=5.17.0

# 数据库（可选）
psycopg2-binary>=2.9.0  # PostgreSQL
redis>=5.0.0

# LLM支持（可选）
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5

# Web服务（可选）
streamlit>=1.29.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.0.0
